requirements.txt
flask == 3.1.0
flask-cors == 5.0.0
Pillow == 11.0.0
#torch == 2.5.1
diffusers == 0.31.0
#torchvision == 0.20.1
numpy == 2.1.3
huggingface_hub == 0.26.3
accelerate == 1.1.1
transformers == 4.46.3
flask-socketio == 5.4.1
#pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118